remove posture checks from service locations wizard
