+++ web/src/pages/AddLocationPage/AddLocationPage.tsx
@@ -87,17 +87,18 @@ export const AddLocationPage = () => {
         label: m.add_location_step_access_control_label(),
         description: m.add_location_step_access_control_description(),
       },
-      firewall: {
-        id: AddLocationPageStep.Firewall,
-        order: 6,
-        label: m.add_location_step_firewall_label(),
-        description: m.add_location_step_firewall_description(),
-      },
       postureCheck: {
         id: AddLocationPageStep.PostureCheck,
-        order: 7,
+        order: 6,
         label: m.add_location_step_posture_check_label(),
+        hidden: locationType === 'service',
         description: m.add_location_step_posture_check_description(),
+      },
+      firewall: {
+        id: AddLocationPageStep.Firewall,
+        order: 7,
+        label: m.add_location_step_firewall_label(),
+        description: m.add_location_step_firewall_description(),
       },
     }),
     [locationType],
