+++ web/src/pages/AddLocationPage/steps/AddLocationFirewallStep.tsx
@@ -1,11 +1,15 @@
-import { useQuery } from '@tanstack/react-query';
+import { useMutation, useQuery } from '@tanstack/react-query';
+import { useNavigate } from '@tanstack/react-router';
+import { cloneDeep, omit } from 'lodash-es';
 import { useCallback, useMemo, useState } from 'react';
 import { m } from '../../../paraglide/messages';
+import api from '../../../shared/api/api';
 import { ActionCard } from '../../../shared/components/ActionCard/ActionCard';
 import { Controls } from '../../../shared/components/Controls/Controls';
 import { WizardCard } from '../../../shared/components/wizard/WizardCard/WizardCard';
 import { externalLink } from '../../../shared/constants';
 import { Button } from '../../../shared/defguard-ui/components/Button/Button';
+import { Checkbox } from '../../../shared/defguard-ui/components/Checkbox/Checkbox';
 import { Divider } from '../../../shared/defguard-ui/components/Divider/Divider';
 import { Radio } from '../../../shared/defguard-ui/components/Radio/Radio';
 import { SizedBox } from '../../../shared/defguard-ui/components/SizedBox/SizedBox';
@@ -13,14 +17,19 @@ import { ThemeSpacing } from '../../../shared/defguard-ui/types';
 import { isPresent } from '../../../shared/defguard-ui/utils/isPresent';
 import { getLicenseInfoQueryOptions } from '../../../shared/query';
 import { canUseBusinessFeature } from '../../../shared/utils/license';
+import { useGatewayWizardStore } from '../../GatewaySetupPage/useGatewayWizardStore';
 import businessFeatureCardImage from '../assets/business-feature-icon.png';
+import actionCardImage from '../assets/gateway-setup-action-card.png';
 import { AddLocationPageStep } from '../types';
 import { useAddLocationStore } from '../useAddLocationStore';
 
 type Choice = 'disable' | 'enabled-allowed' | 'enabled-denied';
 
 export const AddLocationFirewallStep = () => {
+  const locationType = useAddLocationStore((s) => s.locationType);
   const [state, setState] = useState<Choice>('disable');
+  const [showGateway, setShowGateway] = useState(true);
+  const navigate = useNavigate();
 
   const { data: licenseInfo } = useQuery(getLicenseInfoQueryOptions);
   const canUseFeature = useMemo(() => {
@@ -28,6 +37,29 @@ export const AddLocationFirewallStep = () => {
     return canUseBusinessFeature(licenseInfo).result;
   }, [licenseInfo]);
   const firewallLocked = isPresent(canUseFeature) && !canUseFeature;
+
+  const { mutate, isPending } = useMutation({
+    mutationFn: api.location.addLocation,
+    meta: {
+      invalidate: [['network'], ['enterprise_info']],
+    },
+    onSuccess: ({ data }) => {
+      if (showGateway) {
+        useGatewayWizardStore.getState().start({ network_id: data.id });
+        navigate({ to: '/setup-gateway', replace: true }).then(() => {
+          setTimeout(() => {
+            useAddLocationStore.getState().reset();
+          }, 100);
+        });
+      } else {
+        navigate({ to: '/locations', replace: true }).then(() => {
+          setTimeout(() => {
+            useAddLocationStore.getState().reset();
+          }, 100);
+        });
+      }
+    },
+  });
 
   const saveChanges = useCallback((value: Choice) => {
     const enabled = value !== 'disable';
@@ -39,13 +71,16 @@ export const AddLocationFirewallStep = () => {
   }, []);
 
   const handleSubmit = () => {
-    const enabled = state !== 'disable';
-    const allowed = state === 'enabled-allowed';
-    useAddLocationStore.setState({
-      acl_enabled: enabled,
-      acl_default_allow: allowed,
-      activeStep: AddLocationPageStep.PostureCheck,
-    });
+    saveChanges(state);
+    const storageState = cloneDeep(
+      omit(useAddLocationStore.getState(), [
+        'start',
+        'reset',
+        'activeStep',
+        'locationType',
+      ]),
+    );
+    mutate(storageState);
   };
 
   return (
@@ -94,6 +129,20 @@ export const AddLocationFirewallStep = () => {
         text={m.location_firewall_option_default_deny()}
         disabled={firewallLocked}
       />
+      <Divider spacing={ThemeSpacing.Xl2} />
+      <ActionCard
+        imageSrc={actionCardImage}
+        title={m.add_location_firewall_gateway_activation_title()}
+        subtitle={m.add_location_firewall_gateway_activation_subtitle()}
+      >
+        <Checkbox
+          text={m.add_location_firewall_gateway_activation_checkbox()}
+          active={showGateway}
+          onClick={() => {
+            setShowGateway((s) => !s);
+          }}
+        />
+      </ActionCard>
       <Controls>
         <Button
           variant="outlined"
@@ -101,14 +150,18 @@ export const AddLocationFirewallStep = () => {
           onClick={() => {
             saveChanges(state);
             useAddLocationStore.setState({
-              activeStep: AddLocationPageStep.AccessControl,
+              activeStep:
+                locationType === 'service'
+                  ? AddLocationPageStep.AccessControl
+                  : AddLocationPageStep.PostureCheck,
             });
           }}
         />
         <div className="right">
           <Button
-            testId="firewall-continue"
-            text={m.controls_continue()}
+            testId="create-location"
+            text={m.add_location_create_location()}
+            loading={isPending}
             onClick={() => {
               handleSubmit();
             }}
